+++ data-raw/DATASET.R
@@ -179,12 +179,14 @@ startdate <- "2004-01-01"
 
 crude <- c(
   paste0("CL_", sprintf("%0.3d", 1:36), "_Month"),
-  paste0("NG_", sprintf("%0.3d", 1:36), "_Month")
+  paste0("NG_", sprintf("%0.3d", 1:36), "_Month"),
+  paste0("YX_", sprintf("%0.3d", 1:12), "_Month")
 )
 crudecan <- c(paste0("WCW_", sprintf("%0.3d", 1:12), "_Month"),
               paste0("YV_", sprintf("%0.3d", 1:6), "_Month"))
 crudeICE <- c(paste0("BRN_", sprintf("%0.3d", 1:36), "_Month"))
-pdts <- c(paste0("HO_", sprintf("%0.3d", 1:18), "_Month"), paste0("RB_", sprintf("%0.3d", 1:18), "_Month"))
+pdts <- c(paste0("HO_", sprintf("%0.3d", 1:18), "_Month"),
+          paste0("RB_", sprintf("%0.3d", 1:18), "_Month"))
 
 crude <- RTL::getPrices(
   feed = "CME_NymexFutures_EOD_continuous",
@@ -261,6 +263,7 @@ dflong %>% dplyr::filter(grepl("HO",series)) %>% ggplot(aes(x = date, y = value,
 dflong %>% dplyr::filter(grepl("RB",series)) %>% ggplot(aes(x = date, y = value, col = series)) + geom_line()
 dflong %>% dplyr::filter(grepl("WCW",series)) %>% ggplot(aes(x = date, y = value, col = series)) + geom_line()
 dflong %>% dplyr::filter(grepl("YV",series)) %>% ggplot(aes(x = date, y = value, col = series)) + geom_line()
+dflong %>% dplyr::filter(grepl("YX",series)) %>% ggplot(aes(x = date, y = value, col = series)) + geom_line()
 dflong %>% dplyr::filter(grepl("ALI",series)) %>% ggplot(aes(x = date, y = value, col = series)) + geom_line()
 dflong %>% dplyr::filter(grepl("NG",series)) %>% ggplot(aes(x = date, y = value, col = series)) + geom_line()
 dflong %>% dplyr::filter(grepl("AUP",series)) %>% ggplot(aes(x = date, y = value, col = series)) + geom_line()
@@ -917,9 +920,9 @@ usethis::use_data(fizdiffs, overwrite = T)
 
 tradeHubs <-
   dplyr::tibble(
-    lat = c(53.54623, 52.6735, 35.94068, 30.00623, 29.73144),
-    long = c(-113.34684, -111.3075, -96.74536, -93.96882,-95.12680),
-    hub = c("Edmonton", "Hardisty", "Cushing", "Nederland","Mont Belvieu")
+    lat = c(53.54623, 52.6735, 35.94068, 30.00623, 29.73144, 49.28964),
+    long = c(-113.34684, -111.3075, -96.74536, -93.96882,-95.12680, -122.94499),
+    hub = c("Edmonton", "Hardisty", "Cushing", "Nederland","Mont Belvieu", "Burnaby")
   )
 usethis::use_data(tradeHubs, overwrite = T)