+++ src/RcppExports.cpp
@@ -25,6 +25,24 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// rcppOUJ
+NumericMatrix rcppOUJ(NumericMatrix x, NumericMatrix djump, double theta, double mu, double dt, double sigma, double jump_prob, double jump_avesize);
+RcppExport SEXP _RTL_rcppOUJ(SEXP xSEXP, SEXP djumpSEXP, SEXP thetaSEXP, SEXP muSEXP, SEXP dtSEXP, SEXP sigmaSEXP, SEXP jump_probSEXP, SEXP jump_avesizeSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericMatrix >::type x(xSEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type djump(djumpSEXP);
+    Rcpp::traits::input_parameter< double >::type theta(thetaSEXP);
+    Rcpp::traits::input_parameter< double >::type mu(muSEXP);
+    Rcpp::traits::input_parameter< double >::type dt(dtSEXP);
+    Rcpp::traits::input_parameter< double >::type sigma(sigmaSEXP);
+    Rcpp::traits::input_parameter< double >::type jump_prob(jump_probSEXP);
+    Rcpp::traits::input_parameter< double >::type jump_avesize(jump_avesizeSEXP);
+    rcpp_result_gen = Rcpp::wrap(rcppOUJ(x, djump, theta, mu, dt, sigma, jump_prob, jump_avesize));
+    return rcpp_result_gen;
+END_RCPP
+}
 // rcppOUt
 NumericMatrix rcppOUt(NumericMatrix x, double theta, double dt, double sigma);
 RcppExport SEXP _RTL_rcppOUt(SEXP xSEXP, SEXP thetaSEXP, SEXP dtSEXP, SEXP sigmaSEXP) {
@@ -42,6 +60,7 @@ END_RCPP
 
 static const R_CallMethodDef CallEntries[] = {
     {"_RTL_rcppOU", (DL_FUNC) &_RTL_rcppOU, 5},
+    {"_RTL_rcppOUJ", (DL_FUNC) &_RTL_rcppOUJ, 8},
     {"_RTL_rcppOUt", (DL_FUNC) &_RTL_rcppOUt, 4},
     {NULL, NULL, 0}
 };